assignments/classification/requirements.txt
visdom==0.1.8.3
scipy==1.1.0
torch==0.4.0
numpy==1.14.2
torchvision==0.2.1
Pillow==5.2.0
ignite==1.1.0
tensorboardX==1.2
PyYAML==3.13
